TCF Debugger: Fix Breakpoint Scope Page
Fixed the NullPointException thrown when a reported context cannot be found in the children retrieved via the RunControl service.

When selecting the Scope of a breakpoint from Breakpoint Properties, if a reported thread does not belog to the list of children retrieved using the RunControl service, two NullPointerExceptions are thrown and the Scope page won't be opened.

The current Merge Request is a proposal for solving this issue within the TCF code: eclipse/tcf/tcf#107 (https://gitlab.eclipse.org/eclipse/tcf/tcf/-/issues/107)